# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Academic World Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mysql/top-keywords:
    get:
      operationId: mysqlTopKeywords
      tags: [mysql]
      summary: Top keywords by publication count
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KeywordCount"
  /mysql/faculty-search:
    get:
      operationId: mysqlFacultySearch
      tags: [mysql]
      summary: Search faculty by name
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacultyResult"
  /mysql/faculty-publications:
    get:
      operationId: mysqlFacultyPublications
      tags: [mysql]
      summary: Get publications for a faculty member
      parameters:
        - in: query
          name: facultyId
          required: true
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicationResult"
  /mysql/keyword-trend:
    get:
      operationId: mysqlKeywordTrend
      tags: [mysql]
      summary: Publication count per year for a keyword
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YearCount"
  /mysql/top-universities:
    get:
      operationId: mysqlTopUniversities
      tags: [mysql]
      summary: Top universities by publication count
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UniversityCount"
  /mysql/faculty-hindex:
    get:
      operationId: mysqlFacultyHindex
      tags: [mysql]
      summary: Faculty ranked by H-index (window functions + CTE)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacultyHindex"
  /mysql/keyword-cooccurrence:
    get:
      operationId: mysqlKeywordCooccurrence
      tags: [mysql]
      summary: Keywords that frequently co-occur (self-join)
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CooccurrenceResult"
  /mysql/university-profile:
    get:
      operationId: mysqlUniversityProfile
      tags: [mysql]
      summary: University research profile (multi-level aggregation)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UniversityProfile"
  /mysql/favorite-keywords:
    get:
      operationId: mysqlGetFavoriteKeywords
      tags: [mysql]
      summary: Get user favorite keywords
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoriteKeyword"
    post:
      operationId: mysqlAddFavoriteKeyword
      tags: [mysql]
      summary: Add favorite keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteKeywordRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteKeyword"
  /mysql/favorite-keywords/{id}:
    delete:
      operationId: mysqlDeleteFavoriteKeyword
      tags: [mysql]
      summary: Remove favorite keyword
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  # ── MongoDB endpoints ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    KeywordCount:
      type: object
      properties:
        keyword:
          type: string
        count:
          type: integer
      required: [keyword, count]
...
    UniversityCount:
      type: object
      properties:
        university:
          type: string
        count:
          type: integer
      required: [university, count]
    UniversityProfile:
      type: object
      properties:
        university:
          type: string
        facultyCount:
          type: integer
        totalPublications:
          type: integer
        avgCitations:
          type: number
        keywordDiversity:
          type: integer
        topKeyword:
          type: string
      required: [university, facultyCount, totalPublications, avgCitations, keywordDiversity]
    FacultyResult:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: string
        university:
          type: string
        email:
          type: string
        photoUrl:
          type: string
      required: [id, name]
    FacultyHindex:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        university:
          type: string
        hIndex:
          type: integer
        totalPublications:
          type: integer
        totalCitations:
          type: integer
      required: [id, name, hIndex, totalPublications, totalCitations]
    CooccurrenceResult:
      type: object
      properties:
        keyword:
          type: string
        cooccurrences:
          type: integer
      required: [keyword, cooccurrences]
    PublicationResult:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        year:
          type: integer
        numCitations:
          type: integer
        venue:
          type: string
      required: [id, title]
...
    YearCount:
      type: object
      properties:
        year:
          type: integer
        count:
          type: integer
      required: [year, count]
    FavoriteKeyword:
      type: object
      properties:
        id:
          type: integer
        keyword:
          type: string
        createdAt:
          type: string
      required: [id, keyword]
    AddFavoriteKeywordRequest:
      type: object
      properties:
        keyword:
          type: string
      required: [keyword]
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required: [success]